backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.20.0
supabase==1.0.3
pydantic==1.10.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0